ai-service/requirements.txt
fastapi
uvicorn[standard]
transformers
torch>=2.11.0
pydantic
python-dotenv
httpx
